Asana, what does it mean anyway?
Asana—have you noticed how asana follows all the descriptive names of a yoga pose; e.g., Sarvangasana, Shavasana, Tadasana and so on? Asana literally means seat, or more precisely stated in the Yoga Sutras (II.46), it is elaborated as a sturdy but yet, a relaxed and comfortable pose. Patanjali notesYoga Sutras (II.46), Sthira; translation is steady and comfortable, Sukham translation is easy, blissful, prosperous, easiness, enjoyably, asanam; translation is posture, presence, sitting still without disruption. Patanjali says that a correct posture must be strong without tension so the person can sit for meditation as well as do the physical tasks and duties in life (II.46).

So, what can we do with Patanjali’s concept of asana in our daily yoga practice?

One can take use of these notations from the Yoga Sutras which so beautifully remind us that we are not to push ourselves in a posture. Our ego wants us to go as far as we can in a pose in the shortest time possible, and although we keep this secret, the ego also wants us to be better than everyone else. So not only do we have the responsibility to keep our body in check, but we also need to do the same with our ego. To do this we not merely practicing meditation, but we also keep in mind the meaning of Asana. Asana with its purpose reminds us to listen to our body and not go too far in a pose, but instead back of a little until we are in a steady, comfortable posture. The meaning of Asana is perhaps a strange and foreign concept to us here in the West because we are taught to do our best and to push ourselves in life, especially at the gym, but yoga, or a asana practice, is not like any other workout; it can be a physical workout, yes, but it is also mental, and a spiritual workout if you will. When we are in a steady, comfortable pose, it is then that yoga can do its magic. The energies known as prana(s) can flow freely throughout our body since there are no restrictions. If we keep this in mind, we will begin to get a little closer to the meaning of yoga, and we will also start to experience how a yoga practice can transform our lives.

Author: Tone Jackson is a certified yoga instructor and a licensed massage therapist. She is the owner of Healthwork Yoga and Massage Therapy in Albuquerque, NM.

References:

Sri Swami Satchidananda. (1978). The Yoga Sutras of Patanjali. Buckingham, VA: Integral Yoga Publications.